Rename module: `use_surface_geo_mask` -> `masks`
