+++ seaice_ecdr/complete_daily_ecdr.py
@@ -19,6 +19,9 @@ from pm_tb_data._types import NORTH, SOUTH, Hemisphere
 from seaice_ecdr._types import ECDR_SUPPORTED_RESOLUTIONS
 from seaice_ecdr.cli.util import datetime_to_date
 from seaice_ecdr.constants import STANDARD_BASE_OUTPUT_DIR
+from seaice_ecdr.masks import (
+    get_surfacetype_da,
+)
 from seaice_ecdr.melt import (
     MELT_ONSET_FILL_VALUE,
     MELT_SEASON_FIRST_DOY,
@@ -27,9 +30,6 @@ from seaice_ecdr.melt import (
 )
 from seaice_ecdr.set_daily_ncattrs import finalize_cdecdr_ds
 from seaice_ecdr.temporal_composite_daily import get_tie_filepath, make_tiecdr_netcdf
-from seaice_ecdr.use_surface_geo_mask import (
-    get_surfacetype_da,
-)
 from seaice_ecdr.util import standard_daily_filename
