+++ seaice_ecdr/masks.py
